+++ packages/server/migrations/006_session_other_info.up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE session_entities ADD COLUMN current_word_other_info TEXT;